+++ ReleaseNotes.md
@@ -1,2 +1,10 @@
+### (2021-02-17)
+
+##### MpSoft.AspNet.Grpc.MvcApiGenerator - 1.0.1
+* add Authorize and AllowAnonymous attributes support
+
+##### MpSoft.AspNet.Grpc.MvcApiGenerator.Runtime - 1.0.1
+* improve HttpContext propagation
+
 ### 1.0.0 (2020-12-18)  
 Initial version